Added write to error topic and bumped dependencies

+++ src/main/java/de/uksh/medic/etl/settings/CxxMdrSettings.java
@@ -24,7 +24,7 @@ public class CxxMdrSettings {
 
     public void setToken(String newToken, int expiresIn) {
         this.token = newToken;
-        tokenExpiration = System.currentTimeMillis() + SECONDMILLISECOND * expiresIn;
+        tokenExpiration = System.currentTimeMillis() + (long) SECONDMILLISECOND * expiresIn;
     }
 
     public String getUsername() {
